+++ src/lib/counter.test.ts
@@ -1,24 +1,24 @@
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
 
-import {toCounter} from './counter.js';
+import {create_counter} from './counter.js';
 
-/* test__toCounter */
-const test__toCounter = suite('toCounter');
+/* test__create_counter */
+const test__create_counter = suite('create_counter');
 
-test__toCounter('basic behavior', () => {
-	const counter = toCounter();
+test__create_counter('basic behavior', () => {
+	const counter = create_counter();
 	assert.is(counter(), 0);
 	assert.is(counter(), 1);
 	assert.is(counter(), 2);
 });
 
-test__toCounter('custom count', () => {
-	const counter = toCounter(1);
+test__create_counter('custom count', () => {
+	const counter = create_counter(1);
 	assert.is(counter(), 1);
 	assert.is(counter(), 2);
 	assert.is(counter(), 3);
 });
 
-test__toCounter.run();
-/* test__toCounter */
+test__create_counter.run();
+/* test__create_counter */