+++ src/lib/counter.ts
@@ -2,12 +2,12 @@ export interface Counter {
 	(): number;
 }
 
-export interface ToCounter {
+export interface create_counter {
 	(initial?: number): Counter;
 }
 
 // 0-based counter by default
-export const toCounter: ToCounter = (count = 0) => {
+export const create_counter: create_counter = (count = 0) => {
 	let c = count;
 	return () => c++;
 };