+++ src/lib/json.ts
@@ -2,7 +2,7 @@ export type Json = string | number | boolean | null | {[prop: string]: Json} | J
 
 export type JsonType = 'string' | 'number' | 'boolean' | 'null' | 'object' | 'array';
 
-export const get_json_type = (value: Json): JsonType | undefined => {
+export const to_json_type = (value: Json): JsonType | undefined => {
 	const type = typeof value;
 	switch (type) {
 		case 'string':